KStreams tutorial run instructions (#109)
* chore: add run instructions to KStreams splitting tutorial

* chore: add run instructions to KStreams merging tutorial

* chore: add run instructions to KStreams dynamic output topic tutorial

* chore: add run instructions to KStreams filtering tutorial

* chore: add run instructions to KStreams windowed dedupe tutorial

* chore: add run instructions to KStreams first app tutorial

* chore: add run instructions to KStreams stream to table conversion tutorial

* chore: add run instructions to KStreams transforming tutorial

* chore: add run instructions to KStreams count aggregation tutorial

* chore: add run instructions to KStreams sum aggregation tutorial

* chore: add run instructions to KStreams min/max aggregation tutorial

* chore: add run instructions to KStreams cogrouping tutorial

* chore: add run instructions to KStreams stream-table join tutorial

* chore: add run instructions to KStreams stream-stream join tutorial

* chore: add run instructions to KStreams table-table join tutorial

* chore: add run instructions to KStreams table-table FK join tutorial

* chore: add run instructions to KStreams versioned KTables tutorial

* chore: add run instructions to KStreams tumbling windows tutorial

* chore: add run instructions to KStreams sliding windows tutorial

* chore: add run instructions to KStreams session windows tutorial

* chore: add run instructions to KStreams window final result emission tutorial

* chore: add run instructions to KStreams scheduling operations tutorial

* chore: add run instructions to KStreams KTable TTL tutorial

* chore: add run instructions to KStreams KTable aggregate TTL tutorial

* chore: add run instructions to KStreams uncaught exception handling tutorial

* fix: typo in KStreams tutorial Gradle command

+++ kafka-streams-schedule-operations/kstreams/src/main/java/io/confluent/developer/KafkaStreamsPunctuation.java
@@ -43,29 +43,29 @@ public class KafkaStreamsPunctuation {
                 .peek((key, value) -> LOG.info("Incoming records key[{}] value[{}]", key, value))
                 .process(new PunctationProcessorSupplier(), Named.as("max-login-time-transformer"), LOGIN_TIME_STORE)
                 .peek((key, value) -> LOG.info("Punctuation records key[{}] value[{}]", key, value))
-                      .to(OUTPUT_TOPIC, Produced.with(Serdes.String(), Serdes.Long()));
+                      .to(OUTPUT_TOPIC, Produced.with(Serdes.String(), Serdes.Integer()));
 
         return builder.build(envProps);
     }
 
 
-    private static class PunctationProcessorSupplier implements ProcessorSupplier<String,LoginTime, String, Long> {
+    private static class PunctationProcessorSupplier implements ProcessorSupplier<String, LoginTime, String, Integer> {
 
         @Override
         public Set<StoreBuilder<?>> stores() {
-            StoreBuilder<KeyValueStore<String, Long>> storeBuilder =
-                    Stores.keyValueStoreBuilder(Stores.inMemoryKeyValueStore(LOGIN_TIME_STORE),Serdes.String(), Serdes.Long());
+            StoreBuilder<KeyValueStore<String, Integer>> storeBuilder =
+                    Stores.keyValueStoreBuilder(Stores.inMemoryKeyValueStore(LOGIN_TIME_STORE),Serdes.String(), Serdes.Integer());
             return Collections.singleton(storeBuilder);
         }
 
         @Override
-        public Processor<String, LoginTime, String, Long> get() {
-            return new Processor<String, LoginTime, String, Long>() {
-                private KeyValueStore<String, Long> store;
-                private ProcessorContext<String, Long> context;
+        public Processor<String, LoginTime, String, Integer> get() {
+            return new Processor<String, LoginTime, String, Integer>() {
+                private KeyValueStore<String, Integer> store;
+                private ProcessorContext<String, Integer> context;
 
                 @Override
-                public void init(ProcessorContext<String, Long> context) {
+                public void init(ProcessorContext<String, Integer> context) {
                     this.context = context;
                     store = this.context.getStateStore(LOGIN_TIME_STORE);
                     this.context.schedule(Duration.ofSeconds(5), PunctuationType.STREAM_TIME, this::streamTimePunctuator);
@@ -76,28 +76,28 @@ public class KafkaStreamsPunctuation {
                 public void process(Record<String, LoginTime> loginRecord) {
                     String key = loginRecord.key();
                     LoginTime value = loginRecord.value();
-                    Long currentVT = store.putIfAbsent(key, value.logInTime());
+                    Integer currentVT = store.putIfAbsent(key, value.logInTime());
                     if (currentVT != null) {
                         store.put(key, currentVT + value.logInTime());
                     }
                 }
 
                 void wallClockTimePunctuator(Long timestamp) {
-                    try (KeyValueIterator<String, Long> iterator = store.all()) {
+                    try (KeyValueIterator<String, Integer> iterator = store.all()) {
                         while (iterator.hasNext()) {
-                            KeyValue<String, Long> keyValue = iterator.next();
-                            store.put(keyValue.key, 0L);
+                            KeyValue<String, Integer> keyValue = iterator.next();
+                            store.put(keyValue.key, 0);
                         }
                     }
                     System.out.println("@" + new Date(timestamp) + " Reset all view-times to zero");
                 }
 
                 void streamTimePunctuator(Long timestamp) {
-                    Long maxValue = Long.MIN_VALUE;
+                    Integer maxValue = Integer.MIN_VALUE;
                     String maxValueKey = "";
-                    try (KeyValueIterator<String, Long> iterator = store.all()) {
+                    try (KeyValueIterator<String, Integer> iterator = store.all()) {
                         while (iterator.hasNext()) {
-                            KeyValue<String, Long> keyValue = iterator.next();
+                            KeyValue<String, Integer> keyValue = iterator.next();
                             if (keyValue.value > maxValue) {
                                 maxValue = keyValue.value;
                                 maxValueKey = keyValue.key;
